Add a docker-compose example
The example deploys:
   - TimescaleDB
   - Timescale-Prometheus Connector
   - Prometheus
   - cAdvisor
   - Grafana

cAdvisor is used to collect container metrics and is then
scraped by Prometheus. Timescale-Prometheus is configured
as remote storage for Prometheus and stores the data in
TimescaleDB. Grafana is configured to use TimescaleDB as a
database, but also to have TimescaleDB and Prometheus as
provisioned data sources.